chore: change type label; resolve warnings

+++ Equivnox/src/Math/MathAlgorithm.cpp
@@ -121,9 +121,9 @@ namespace EQX
 		if (x >= 2)
 			return 0;
 		else if (x <= 1)
-			return 1.f / 6.f * (-3 * std::pow(1 - x, 3) + 3 * std::pow(1 - x, 2) + 3 * (1 - x) + 1);
+			return 1.f / 6.f * (-3 * std::powf(1 - x, 3) + 3 * std::powf(1 - x, 2) + 3 * (1 - x) + 1);
 		else
-			return 1.f / 6.f * std::pow(2 - x, 3);
+			return 1.f / 6.f * std::powf(2 - x, 3);
 	}
 
 	float EvalCRCubic(float x)
@@ -133,9 +133,9 @@ namespace EQX
 		if (x >= 2)
 			return 0;
 		if (x <= 1)
-			return 1.f / 2.f * (-3 * std::pow(1 - x, 3) + 4 * std::pow(1 - x, 2) + (1 - x));
+			return 1.f / 2.f * (-3 * std::powf(1 - x, 3) + 4 * std::powf(1 - x, 2) + (1 - x));
 		else
-			return 1.f / 2.f * (std::pow(2 - x, 3) - std::pow(2 - x, 2));
+			return 1.f / 2.f * (std::powf(2 - x, 3) - std::powf(2 - x, 2));
 	}
 
 }
